Refactoring - minor fix - 3
